requirements.txt
pdfminer.six
xmltodict
